Minor refactoring and improvements.
Added printing directly to this module so it doesn't have to
implicitly depend on another module.

The Makefile now makes use of a list of -Wl,-U symbols so that it
can build this and expect to find those symbols provided by the host.

Lots of changes to the oscript.ose file.